reformat data table viewer

+++ src/file/row_groups.rs
@@ -34,7 +34,7 @@ pub struct RowGroupColumnStats {
 }
 
 pub struct RowGroupColumnMetadata {
-    pub file_offset: i64,
+    pub file_offset: u64,
     pub column_path: String,
     pub has_stats: HasStats,
     pub statistics: Option<RowGroupColumnStats>,
@@ -181,7 +181,7 @@ impl RowGroupColumnMetadata {
         let statistics = RowGroupColumnStats::new(column_chunk.statistics());
 
         Ok(RowGroupColumnMetadata {
-            file_offset: column_chunk.file_offset(),
+            file_offset: column_chunk.file_offset() as u64,
             column_path: column_chunk.column_descr().path().to_string(),
             has_stats: HasStats {
                 has_stats: column_chunk.statistics().is_some(),
